+++ Api/Glass.API.Backend/Models/Clientes/V1/Filtro/ComissionadoDto.cs
@@ -19,6 +19,6 @@ namespace Glass.API.Backend.Models.Clientes.V1.Filtro
         /// </summary>
         [DataMember]
         [JsonProperty("percentual")]
-        public double Percentual { get; set; }
+        public decimal Percentual { get; set; }
     }
 }
